+++ api/v1/warehouse/entity.go
@@ -24,11 +24,11 @@ type Location struct {
 	Level          string    `db:"level" json:"level"`
 	BayID          string    `db:"bay_id" json:"bay_id"`
 	ItemID         string    `db:"item_id" json:"item_id"`
-	Capacity       int64     `db:"capacity" json:"capacity"`
-	Quantity       int64     `db:"quantity" json:"quantity"`
-	Available      int64     `db:"available" json:"available"`
-	CanPick        int64     `db:"can_pick" json:"can_pick"`
-	Alert          int64     `db:"alert" json:"alert"`
+	Capacity       int       `db:"capacity" json:"capacity"`
+	Quantity       int       `db:"quantity" json:"quantity"`
+	Available      int       `db:"available" json:"available"`
+	CanPick        int       `db:"can_pick" json:"can_pick"`
+	Alert          int       `db:"alert" json:"alert"`
 	Status         int       `db:"status" json:"status"`
 	Created        time.Time `db:"created" json:"created"`
 	CreatedBy      string    `db:"created_by" json:"created_by"`
